Moved Events folder out of Models folder
